Restructure the project. Create solvex for package and api for FastAPI app
